+++ packages/video_player_avplay/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.7.7
+
+* Fix black line issue when playing video.
+
+## 0.7.6
+* Fix video doesn't scale issue.
+
 ## 0.7.5
 * Fix the issue of subtitles remaining longer than their end time.